refactor: share legacy log parser
