+++ src/progressView/events/ProgressEventHandler.ts
@@ -1,7 +1,5 @@
 // Third-party imports
 import * as vscode from 'vscode';
-import { decode as decodeHtml } from 'he';
-
 // Local imports
 import { ProgressViewState } from '../state/ProgressViewState';
 import { WebviewUpdater } from '../managers';
@@ -13,7 +11,7 @@ import { onProgress } from '@eventBus/ProgressEventBus';
 import { TokenUsageStats } from '@agent/types/UsageTypes';
 import { TaskState } from '@logger/TaskState';
 import { LogMessageData, TaskGroup } from '@logger/LogTypes';
-import { MESSAGE_TYPES } from '@logger/messageTypes';
+import { parseLegacyLogData } from '@logger/logUtils';
 import type { StreamTabId, ExecutionId } from '@agent/types/IdentifierTypes';
 
 // @ts-ignore - Import JavaScript module
@@ -52,43 +50,6 @@ export class ProgressEventHandler {
    * missing. Parsed results are stored back into the log object so that future
    * lookups don't require re-parsing.
    */
-  private parseLegacyData(
-    logMessage: LogMessageData,
-    forceParse = false,
-  ): unknown | undefined {
-    if (!forceParse && logMessage.data !== undefined) {
-      return logMessage.data;
-    }
-
-    const type = logMessage.messageType;
-    const legacyTypes = new Set<string>([
-      MESSAGE_TYPES.FILE_LIST,
-      MESSAGE_TYPES.MISSING_OUTPUTS,
-      MESSAGE_TYPES.LATEXDIFF,
-      MESSAGE_TYPES.STATISTICS,
-    ]);
-
-    if (type && legacyTypes.has(type) && logMessage.text) {
-      try {
-        const decoded = decodeHtml(logMessage.text);
-        const parsed = JSON.parse(decoded);
-
-        // Sanity check that parsed value is an object or array
-        if (typeof parsed === 'object' && parsed !== null) {
-          logMessage.data = parsed;
-          return parsed;
-        }
-      } catch (err) {
-        this.logger.warn(
-          `Failed to parse legacy log data: ${
-            err instanceof Error ? err.message : String(err)
-          }`,
-        );
-      }
-    }
-    return undefined;
-  }
-
   /**
    * Setup all event bus listeners
    */
@@ -340,7 +301,7 @@ export class ProgressEventHandler {
   }): void {
     const { stream, logMessage } = data;
 
-    this.parseLegacyData(logMessage);
+    parseLegacyLogData(logMessage, this.logger);
 
     // Skip debug messages if debug mode is disabled
     if (
@@ -382,7 +343,7 @@ export class ProgressEventHandler {
       existing.data = logMessage.data;
     } else {
       // Re-parse the updated text even if existing data is present
-      this.parseLegacyData(existing, true);
+      parseLegacyLogData(existing, this.logger, true);
     }
 
     if (
